Simple preemptive multi-processing

+++ src/lib/panic.h
@@ -5,7 +5,7 @@
 #include "../interrupt/cpu_state.h"
 #include "../interrupt/handlers.h"
 
-uint8_t helper_variable;
+uint32_t helper_variable;
 
 void PANIC(char* msg);
 void PANIC_EXC(char* msg, exception_state* state);
